J-6 before paper

+++ src/dnfpy/model/activationMap.py
@@ -2,9 +2,10 @@ import numpy as np
 from dnfpy.core.map2D import Map2D
 
 class ActivationMap(Map2D):
-    def __init__(self,name,size,dt=0.1,model='cnft',th=0.75,dtype=np.float32,**kwargs):
+    def __init__(self,name,size,dt=0.1,model='cnft',th=0.75,**kwargs):
         super(ActivationMap,self).__init__(name,
-            size=size,dt=dt,model=model,th=th,dtype=dtype,**kwargs)
+            size=size,dt=dt,model=model,th=th,**kwargs)
+
     def _compute(self,model,field,th,dtype):
         if model == 'cnft':
             self._data = np.maximum(field,dtype(0))
